junk/do-search/pyproject.toml
[project]
name = "do-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "lxml>=6.0.2",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
